Sanctum Authentication

+++ routes/api.php
@@ -21,38 +21,36 @@ use App\Http\Controllers\PromotionController;
 */
 
 
-Route::middleware('auth:api')->group(function()
-{
-    Route::get('/users/me', [AuthController::class,'userInfo']);
-});
+Route::get('/users/me', [AuthController::class,'userInfo'])->middleware('auth:sanctum');
 // make route resources              ##
 
 Route::post('/users/register', [AuthController::class,'register']);
 Route::post('/users/login', [AuthController::class,'login']);
+Route::get('/users/logout',[AuthController::class,'logout'])->middleware('auth:sanctum');
 Route::get('/users',[UserController::class,'index']);
 Route::get('/users/{user}',[UserController::class,'show']);
 
 Route::get('/articles',[ArticleController::class,'index']);
-Route::post('/articles',[ArticleController::class,'store']);
+Route::post('/articles',[ArticleController::class,'store'])->middleware('auth:sanctum');
 Route::get('/articles/{article}',[ArticleController::class,'show']);
-Route::put('/articles/{article}',[ArticleController::class,'update']); //change from form-data to x-www-form-urlencoded 
-Route::delete('/articles/{article}',[ArticleController::class,'destroy']);
+Route::put('/articles/{article}',[ArticleController::class,'update'])->middleware('auth:sanctum'); //change from form-data to x-www-form-urlencoded 
+Route::delete('/articles/{article}',[ArticleController::class,'destroy'])->middleware('auth:sanctum');
 
 Route::get('/questions',[QuestionController::class,'index']);
-Route::post('/questions',[QuestionController::class,'store']);
+Route::post('/questions',[QuestionController::class,'store'])->middleware('auth:sanctum');
 Route::get('/questions/{question}',[QuestionController::class,'show']);
-Route::put('/questions/{question}',[QuestionController::class,'update']); 
-Route::delete('/questions/{question}',[QuestionController::class,'destroy']);
+Route::put('/questions/{question}',[QuestionController::class,'update'])->middleware('auth:sanctum'); 
+Route::delete('/questions/{question}',[QuestionController::class,'destroy'])->middleware('auth:sanctum');
 
 Route::get('/events',[EventController::class,'index']);
-Route::post('/events',[EventController::class,'store']);
+Route::post('/events',[EventController::class,'store'])->middleware('auth:sanctum');
 Route::get('/events/{event}',[EventController::class,'show']);
-Route::put('/events/{event}',[EventController::class,'update']);  
-Route::delete('/events/{event}',[EventController::class,'destroy']);
+Route::put('/events/{event}',[EventController::class,'update'])->middleware('auth:sanctum');  
+Route::delete('/events/{event}',[EventController::class,'destroy'])->middleware('auth:sanctum');
 
-Route::post('/rate',[RateController::class,'addRate']);
+Route::post('/rate',[RateController::class,'addRate'])->middleware('auth:sanctum');
 
-Route::put('/promote',[PromotionController::class,'promote']);
+Route::put('/promote',[PromotionController::class,'promote'])->middleware('auth:sanctum');
